ZOOKEEPER-1416
Generalized this feature in the hopes of making it easier to change in the future. The APIs are no longer specific to persistent/recursive. Now, they are generically "add watch" and there is an additional mode that details the behavior of the watch. For now, the only choices are persistent or recursive but that can change in the future.